Update ppp_reconnect.sh
On my R216-Z, the connection command immediately returns "{"ppp_status":"ppp_disconnected"}", but the modem doesn't actually connect. 
I have implemented an additional check that verifies if the modem really changes its state. If it doesn't change, it resets the variable EPOCH so that it can retry immediately after until it change.